requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
httpx==0.27.2
python-dotenv==1.0.1

